feat hard work...

+++ Infrastracture/Models/DesignCategoryModel.cs
@@ -4,7 +4,7 @@ namespace Infrastracture.Models
 {
     public class DesignCategoryModel
     {
-        public int DesignCategoryId { get; set; } 
+        public long DesignCategoryId { get; set; } 
         public string DesignCategoryName { get; set; }
         public int DesignCount { get; set; }
